Included Maximise function

+++ src/support/config/hooks.ts
@@ -1,6 +1,5 @@
 import { Before, BeforeAll, AfterAll, After, AfterStep, setDefaultTimeout, ITestCaseHookParameter, Status, formatterHelpers, BeforeStep } from "@cucumber/cucumber";
-import { chromium, Browser, Page } from "@playwright/test";
-import WebBrowser from "../manager/Browser";
+import { chromium, firefox, webkit, Browser, BrowserContext, Page } from "@playwright/test";
 import fse from "fs-extra";
 import UIActions from "../playwright/actions/UIActions";
 import Log from "../logger/Log";
@@ -8,7 +7,21 @@ import Log from "../logger/Log";
 const timeInMin: number = 60 * 1000;
 setDefaultTimeout(Number.parseInt(process.env.TEST_TIMEOUT, 10) * timeInMin);
 let browser: Browser;
+let context: BrowserContext;
 let page1: Page;
+
+// Function to get window.outerWidth and window.outerHeight
+async function getWindowSize(context: BrowserContext) {
+    const page = await context.newPage();
+    const windowSize = await page.evaluate(() => {
+        return {
+            width: window.outerWidth,
+            height: window.outerHeight
+        };
+    });
+    await page.close();
+    return windowSize;
+}
 
 // Connect to BrowserStack and launch the browser
 BeforeAll(async function () {
@@ -61,14 +74,12 @@ Before(async function ({ pickle, gherkinDocument }: ITestCaseHookParameter) {
         acceptDownloads: true,
         recordVideo: process.env.RECORD_VIDEO === "true" ? { dir: './test-results/videos' } : undefined,
     });
-    page1 = await this.context.newPage();
-    await page1.goto("https://opensource-demo.orangehrmlive.com/web/index.php/auth/login");
-    await page1.waitForLoadState('domcontentloaded');
-    await page1.getByPlaceholder('Username').click();
-    await page1.getByPlaceholder('Username').fill("Admin");
-    await page1.getByPlaceholder('Password').fill("admin123");
-    await page1.getByPlaceholder('Password').press('Enter');
 
+    const windowSize = await getWindowSize(this.context);
+    console.log(`Window size: ${windowSize.width}x${windowSize.height}`);
+
+    page1 = await this.context.newPage();
+    
     this.web = new UIActions(page1);
     // this.rest = new RESTRequest(page1);
     // this.soap = new SOAPRequest();
@@ -96,17 +107,19 @@ After(async function ({ result, pickle, gherkinDocument }: ITestCaseHookParamete
     const { line } = formatterHelpers.PickleParser.getPickleLocation({ gherkinDocument, pickle });
     const status = result.status;
     const scenario = pickle.name;
-    const videoPath = await page1.video()?.path();
 
     await page1.close();
     await this.context.close();
 
     if (process.env.RECORD_VIDEO === "true") {
-        if (status === Status.FAILED) {
-            fse.renameSync(videoPath, `./test-results/videos/${scenario}(${line}).webm`);
-            await this.attach(fse.readFileSync(`./test-results/videos/${scenario}(${line}).webm`), 'video/webm');
-        } else {
-            fse.unlinkSync(videoPath);
+        const videoPath = await page1.video()?.path();
+        if (videoPath) {
+            if (status === Status.FAILED) {
+                fse.renameSync(videoPath, `./test-results/videos/${scenario}(${line}).webm`);
+                await this.attach(fse.readFileSync(`./test-results/videos/${scenario}(${line}).webm`), 'video/webm');
+            } else {
+                fse.unlinkSync(videoPath);
+            }
         }
     }
     Log.testEnd(`${scenario}: ${line}`, status);
